backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Limiter==3.4.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.6
python-dotenv==1.0.1
passlib==1.7.4
PyJWT==2.8.0
requests
bcrypt==4.0.1
sentence-transformers==2.7.0
pdfplumber==0.9.0
scikit-learn==1.5.0
numpy==1.26.4
keybert==0.8.0
nltk==3.8.1
torch~=2.2.0
langchain==0.3.27  
langchain-openai==0.3.29  
langchain-google-genai==2.0.10  
langchain-community==0.3.27
google-generativeai==0.8.5 
requests>=2.25.0
google-cloud-texttospeech==2.27.0
pydub>=0.25.0
